chore: create test util
Separate utility from main crate.

- Must document functions and what they do
- Also See if we can improve test utility for fallible
  and not fallible.
- And finally move invalid to test_case structure
