+++ src/lib.rs
@@ -29,13 +29,6 @@ pub struct Arguments {
     pub annotate: bool,
 }
 
-#[cfg(test)]
-mod test_util {
-    // Manual include, otherwise, we have to make this part of the interface to make use of these
-    // utility functions in tests. We don't want this to be part of the interface.
-    include!("../tests/common.rs");
-}
-
 /// Convert `*.mamba` files to `*.py`.
 ///
 /// For input, the rules are as follows:
