+++ src/parse/control_flow_expr.rs
@@ -110,7 +110,7 @@ mod test {
     use crate::parse::ast::{Node, AST};
     use crate::parse::parse_direct;
     use crate::parse::result::ParseResult;
-    use crate::test_util::resource_content;
+    use tests_util::resource_content;
 
     #[test]
     fn if_else_verify() {
