+++ src/parse/expr_or_stmt.rs
@@ -61,7 +61,7 @@ mod test {
     use crate::parse::ast::{Node, AST};
     use crate::parse::parse_direct;
     use crate::parse::result::ParseResult;
-    use crate::test_util::resource_content;
+    use tests_util::resource_content;
 
     #[test]
     fn range_verify() {
