--- tests/check/valid/mod.rs
@@ -1,3 +1,0 @@
-pub mod class;
-pub mod definition;
-pub mod function;
